+++ main/pools/pools_presets.js
@@ -1,5 +1,5 @@
 ﻿'use strict';
-//01/10/24
+//03/11/24
 
 /* exported createPoolPresets */
 
@@ -7,8 +7,10 @@ include('..\\..\\helpers\\helpers_xxx.js');
 /* global globTags:readable, globQuery:readable */
 include('..\\..\\helpers\\helpers_xxx_prototypes.js');
 /* global _qCond:readable */
+include('..\\..\\helpers\\helpers_xxx_tags.js');
+/* global queryJoin:readable, queryCombinations:readable */
 
-function createPoolPresets({size = 50} = {}) {
+function createPoolPresets({ size = 50 } = {}) {
 	const half = Math.floor(size / 2) + Math.ceil(size % 4 / 2);
 	const fourth = Math.floor(size / 4) || 1;
 	const eighth = Math.floor(size / 8) || 1;
@@ -26,7 +28,7 @@ function createPoolPresets({size = 50} = {}) {
 					query: {
 						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND ' + globQuery.noInstrumental,
 						_LIBRARY_1: globTags.rating + ' EQUAL 4',
-						_LIBRARY_2: globTags.rating + ' EQUAL 5'
+						_LIBRARY_2: globQuery.ratingTop
 					},
 					toPls: 'Top tracks mix',
 					sort: '', // Random
@@ -43,7 +45,7 @@ function createPoolPresets({size = 50} = {}) {
 					query: {
 						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND ' + globQuery.noInstrumental,
 						_LIBRARY_1: globTags.rating + ' EQUAL 4',
-						_LIBRARY_2: globTags.rating + ' EQUAL 5'
+						_LIBRARY_2: globQuery.ratingTop
 					},
 					toPls: 'Top tracks mix',
 					harmonicMix: true
@@ -60,7 +62,7 @@ function createPoolPresets({size = 50} = {}) {
 					query: {
 						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND ' + globQuery.noInstrumental,
 						_LIBRARY_1: globTags.rating + ' EQUAL 4',
-						_LIBRARY_2: globTags.rating + ' EQUAL 5'
+						_LIBRARY_2: globQuery.ratingTop
 					},
 					insertMethod: 'intercalate',
 					toPls: 'Top tracks mix'
@@ -78,7 +80,7 @@ function createPoolPresets({size = 50} = {}) {
 					query: {
 						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)#',
 						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)#',
-						_LIBRARY_2: globTags.rating + ' EQUAL 5 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)#'
+						_LIBRARY_2: globQuery.ratingTop + ' AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)#'
 					},
 					toPls: 'Top last year tracks mix',
 					sort: '',
@@ -95,7 +97,7 @@ function createPoolPresets({size = 50} = {}) {
 					query: {
 						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)#',
 						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)#',
-						_LIBRARY_2: globTags.rating + ' EQUAL 5 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)#'
+						_LIBRARY_2: globQuery.ratingTop + ' AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)#'
 					},
 					toPls: 'Top last 5 years tracks mix',
 					sort: '',
@@ -112,7 +114,7 @@ function createPoolPresets({size = 50} = {}) {
 					query: {
 						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,10)#',
 						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,10)#',
-						_LIBRARY_2: globTags.rating + ' EQUAL 5 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,10)#'
+						_LIBRARY_2: globQuery.ratingTop + ' AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,10)#'
 					},
 					toPls: 'Top last 10 years tracks mix',
 					sort: '',
@@ -128,9 +130,9 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_2: half
 					},
 					query: {
-						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND %ADDED% DURING LAST 5 WEEKS',
-						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND %ADDED% DURING LAST 5 WEEKS',
-						_LIBRARY_2: globTags.rating + ' EQUAL 5 AND %ADDED% DURING LAST 5 WEEKS'
+						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND %ADDED% DURING LAST 4 WEEKS',
+						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND %ADDED% DURING LAST 4 WEEKS',
+						_LIBRARY_2: globQuery.ratingTop + ' AND %ADDED% DURING LAST 4 WEEKS'
 					},
 					toPls: 'Top recently added tracks mix',
 					sort: '',
@@ -147,9 +149,9 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_2: half
 					},
 					query: {
-						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND (%LAST_PLAYED_ENHANCED% DURING LAST 5 WEEKS OR %LAST_PLAYED% DURING LAST 5 WEEKS)',
-						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND (%LAST_PLAYED_ENHANCED% DURING LAST 5 WEEKS OR %LAST_PLAYED% DURING LAST 5 WEEKS)',
-						_LIBRARY_2: globTags.rating + ' EQUAL 5 AND (%LAST_PLAYED_ENHANCED% DURING LAST 5 WEEKS OR %LAST_PLAYED% DURING LAST 5 WEEKS)'
+						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND ' + globQuery.recent,
+						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND ' + globQuery.recent,
+						_LIBRARY_2: globQuery.ratingTop + ' AND ' + globQuery.recent
 					},
 					toPls: 'Top recently played tracks mix',
 					sort: '',
@@ -165,9 +167,9 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_2: half
 					},
 					query: {
-						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND NOT (%LAST_PLAYED_ENHANCED% DURING LAST 2 WEEKS OR %LAST_PLAYED% DURING LAST 2 WEEKS)',
-						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND NOT (%LAST_PLAYED_ENHANCED% DURING LAST 2 WEEK OR %LAST_PLAYED% DURING LAST 2 WEEK)',
-						_LIBRARY_2: globTags.rating + ' EQUAL 5 AND NOT (%LAST_PLAYED_ENHANCED% DURING LAST 2 WEEKS OR %LAST_PLAYED% DURING LAST 2 WEEKS)'
+						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND NOT ' + globQuery.recent,
+						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND NOT ' + globQuery.recent,
+						_LIBRARY_2: globQuery.ratingTop + ' AND NOT ' + globQuery.recent
 					},
 					toPls: 'Top not played tracks mix',
 					sort: '',
@@ -182,9 +184,9 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_2: half
 					},
 					query: {
-						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND NOT (%LAST_PLAYED_ENHANCED% DURING LAST 2 WEEKS OR %LAST_PLAYED% DURING LAST 2 WEEKS) AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)#',
-						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND NOT (%LAST_PLAYED_ENHANCED% DURING LAST 2 WEEKS OR %LAST_PLAYED% DURING LAST 2 WEEKS) AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)#',
-						_LIBRARY_2: globTags.rating + ' EQUAL 5 AND NOT (%LAST_PLAYED_ENHANCED% DURING LAST 2 WEEKS OR %LAST_PLAYED% DURING LAST 2 WEEKS) AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)#'
+						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND NOT ' + globQuery.recent + ' AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)#',
+						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND NOT ' + globQuery.recent + ' AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)#',
+						_LIBRARY_2: globQuery.ratingTop + ' AND NOT ' + globQuery.recent + ' AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)#'
 					},
 					toPls: 'Top not played this year tracks mix',
 					sort: '',
@@ -199,9 +201,9 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_2: half
 					},
 					query: {
-						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND NOT (%LAST_PLAYED_ENHANCED% DURING LAST 2 WEEKS OR %LAST_PLAYED% DURING LAST 2 WEEKS) AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)#',
-						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND NOT (%LAST_PLAYED_ENHANCED% DURING LAST 2 WEEKS OR %LAST_PLAYED% DURING LAST 2 WEEKS) AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)#',
-						_LIBRARY_2: globTags.rating + ' EQUAL 5 AND NOT (%LAST_PLAYED_ENHANCED% DURING LAST 2 WEEKS OR %LAST_PLAYED% DURING LAST 2 WEEKS) AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)#'
+						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND NOT ' + globQuery.recent + ' AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)#',
+						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND NOT ' + globQuery.recent + ' AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)#',
+						_LIBRARY_2: globQuery.ratingTop + ' AND NOT ' + globQuery.recent + ' AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)#'
 					},
 					toPls: 'Top not recently played last 5 years tracks mix',
 					sort: '',
@@ -219,7 +221,7 @@ function createPoolPresets({size = 50} = {}) {
 					query: {
 						_LIBRARY_0: globTags.rating + ' EQUAL 3 SORT ASCENDING BY ' + _qCond(globTags.playCount),
 						_LIBRARY_1: globTags.rating + ' EQUAL 4 SORT ASCENDING BY ' + _qCond(globTags.playCount),
-						_LIBRARY_2: globTags.rating + ' EQUAL 5 SORT ASCENDING BY ' + _qCond(globTags.playCount)
+						_LIBRARY_2: globQuery.ratingTop + ' SORT ASCENDING BY ' + _qCond(globTags.playCount)
 					},
 					toPls: 'Top least played tracks mix',
 					sort: '',
@@ -236,7 +238,7 @@ function createPoolPresets({size = 50} = {}) {
 					query: {
 						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)# SORT ASCENDING BY ' + _qCond(globTags.playCount),
 						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)# SORT ASCENDING BY ' + _qCond(globTags.playCount),
-						_LIBRARY_2: globTags.rating + ' EQUAL 5 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)# SORT ASCENDING BY ' + _qCond(globTags.playCount)
+						_LIBRARY_2: globQuery.ratingTop + ' AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,1)# SORT ASCENDING BY ' + _qCond(globTags.playCount)
 					},
 					toPls: 'Top least played this year tracks mix',
 					sort: '',
@@ -253,7 +255,7 @@ function createPoolPresets({size = 50} = {}) {
 					query: {
 						_LIBRARY_0: globTags.rating + ' EQUAL 3 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)# SORT ASCENDING BY ' + _qCond(globTags.playCount),
 						_LIBRARY_1: globTags.rating + ' EQUAL 4 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)# SORT ASCENDING BY ' + _qCond(globTags.playCount),
-						_LIBRARY_2: globTags.rating + ' EQUAL 5 AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)# SORT ASCENDING BY ' + _qCond(globTags.playCount)
+						_LIBRARY_2: globQuery.ratingTop + ' AND ' + _qCond(globTags.date) + ' GREATER #$sub(#YEAR#,5)# SORT ASCENDING BY ' + _qCond(globTags.playCount)
 					},
 					toPls: 'Top least played last 5 years tracks mix',
 					sort: '',
@@ -308,7 +310,7 @@ function createPoolPresets({size = 50} = {}) {
 					query: {
 						_LIBRARY_0: '((' + globTags.genre + ' IS #' + globTags.genre + '#) OR (' + globTags.style + ' IS #' + globTags.style + '#)) AND ' + globQuery.notLowRating,
 						_LIBRARY_1: '((' + globTags.genre + ' IS #' + globTags.genre + '#) OR (' + globTags.style + ' IS #' + globTags.style + '#)) AND ' + globQuery.ratingGr3,
-						_LIBRARY_2: '((' + globTags.genre + ' IS #' + globTags.genre + '#) OR (' + globTags.style + ' IS #' + globTags.style + '#)) AND (' + globQuery.ratingTop + ' OR ' + globQuery.loved + ')'
+						_LIBRARY_2: '((' + globTags.genre + ' IS #' + globTags.genre + '#) OR (' + globTags.style + ' IS #' + globTags.style + '#)) AND ' + globQuery.fav
 					},
 					toPls: 'Current genre/style top tracks',
 					sort: '',
@@ -324,7 +326,7 @@ function createPoolPresets({size = 50} = {}) {
 					},
 					query: {
 						_LIBRARY_0: '((' + globTags.genre + ' IS #' + globTags.genre + '#) OR (' + globTags.style + ' IS #' + globTags.style + '#)) AND ' + globQuery.notLowRating,
-						_LIBRARY_1: '((' + globTags.genre + ' IS #' + globTags.genre + '#) OR (' + globTags.style + ' IS #' + globTags.style + '#)) AND (' + globQuery.ratingTop + ' OR ' + globQuery.loved + ')',
+						_LIBRARY_1: '((' + globTags.genre + ' IS #' + globTags.genre + '#) OR (' + globTags.style + ' IS #' + globTags.style + '#)) AND ' + globQuery.fav,
 						_LIBRARY_2: '((' + globTags.genre + ' IS #' + globTags.genre + '#) OR (' + globTags.style + ' IS #' + globTags.style + '#)) AND (' + globQuery.instrumental + ') AND ' + globQuery.notLowRating
 					},
 					toPls: 'Current genre/style and instrumentals',
@@ -342,7 +344,7 @@ function createPoolPresets({size = 50} = {}) {
 					query: {
 						_LIBRARY_0: '((' + globTags.genre + ' IS #' + globTags.genre + '#) OR (' + globTags.style + ' IS #' + globTags.style + '#)) AND NOT (' + globQuery.recent + ') SORT ASCENDING BY ' + _qCond(globTags.playCount),
 						_LIBRARY_1: '((' + globTags.genre + ' IS #' + globTags.genre + '#) OR (' + globTags.style + ' IS #' + globTags.style + '#)) AND ' + globQuery.ratingGr3 + ' SORT ASCENDING BY ' + _qCond(globTags.playCount),
-						_LIBRARY_2: '((' + globTags.genre + ' IS #' + globTags.genre + '#) OR (' + globTags.style + ' IS #' + globTags.style + '#)) AND (' + globQuery.ratingTop + ' OR ' + globQuery.loved + ') SORT ASCENDING BY ' + _qCond(globTags.playCount)
+						_LIBRARY_2: '((' + globTags.genre + ' IS #' + globTags.genre + '#) OR (' + globTags.style + ' IS #' + globTags.style + '#)) AND ' + globQuery.fav + ' SORT ASCENDING BY ' + _qCond(globTags.playCount)
 					},
 					pickMethod: {
 						_LIBRARY_0: 'start',
@@ -363,7 +365,7 @@ function createPoolPresets({size = 50} = {}) {
 					},
 					query: {
 						_LIBRARY_0: globTags.artist + ' IS #' + globTags.artistRaw + '# AND ' + globQuery.ratingGr3,
-						_LIBRARY_1: globTags.artist + ' IS #' + globTags.artistRaw + '# AND (' + globQuery.ratingTop + ' OR ' + globQuery.loved + ')'
+						_LIBRARY_1: globTags.artist + ' IS #' + globTags.artistRaw + '# AND ' + globQuery.fav
 					},
 					toPls: 'Current artist\'s top tracks',
 					sort: '',
@@ -378,7 +380,7 @@ function createPoolPresets({size = 50} = {}) {
 					},
 					query: {
 						_LIBRARY_0: globTags.artist + ' IS #' + globTags.artistRaw + '# AND NOT (' + globQuery.recent + ') SORT ASCENDING BY ' + _qCond(globTags.playCount),
-						_LIBRARY_1: globTags.artist + ' IS #' + globTags.artistRaw + '# AND (' + globQuery.ratingTop + ' OR ' + globQuery.loved + ') SORT ASCENDING BY ' + _qCond(globTags.playCount)
+						_LIBRARY_1: globTags.artist + ' IS #' + globTags.artistRaw + '# AND ' + globQuery.fav + ' SORT ASCENDING BY ' + _qCond(globTags.playCount)
 					},
 					pickMethod: {
 						_LIBRARY_0: 'start',
@@ -483,13 +485,99 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_6: tenth
 					},
 					query: {
-						_LIBRARY_0: '(STYLE IS black metal OR STYLE IS stoner doom OR STYLE IS doom metal OR STYLE IS death metal) AND %RATING% IS 5',
-						_LIBRARY_1: 'STYLE IS black metal AND NOT STYLE IS folk metal AND (%RATING% IS 4 OR %RATING% IS 3)',
-						_LIBRARY_2: 'STYLE IS stoner doom AND NOT STYLE IS folk metal AND (%RATING% IS 4 OR %RATING% IS 3)',
-						_LIBRARY_3: 'STYLE IS doom metal AND NOT STYLE IS folk metal AND (%RATING% IS 4 OR %RATING% IS 3)',
-						_LIBRARY_4: 'STYLE IS folk metal AND (STYLE IS black metal OR STYLE IS stoner doom OR STYLE IS doom metal OR STYLE IS death metal) AND %RATING% GREATER 2',
-						_LIBRARY_5: '(STYLE IS black metal OR STYLE IS stoner doom OR STYLE IS doom metal OR STYLE IS death metal) AND GENRE IS instrumental AND %RATING% GREATER 2',
-						_LIBRARY_6: '(STYLE IS black metal AND STYLE IS ambient metal) OR (STYLE IS atmospheric black metal) OR (STYLE IS death metal AND STYLE IS acoustic) AND %RATING% GREATER 3'
+						_LIBRARY_0: queryJoin([
+							queryJoin(
+								queryCombinations(
+									['black metal', 'stoner doom', 'doom metal', 'death metal'],
+									[globTags.genre, globTags.style],
+									'OR'
+								),
+								'OR'
+							),
+							globQuery.fav
+						],'AND'),
+						_LIBRARY_1: queryJoin([
+							queryJoin([
+								queryJoin(
+									queryCombinations(['black metal'],[globTags.genre, globTags.style],'OR'	),
+									'OR'
+								),
+								queryJoin(
+									queryCombinations(['folk metal'],[globTags.genre, globTags.style],'OR'	),
+									'OR'
+								),
+							], 'AND NOT'),
+							globTags.rating + ' IS 4 OR ' + globTags.rating + ' IS 3'
+						],'AND'),
+						_LIBRARY_2: queryJoin([
+							queryJoin([
+								queryJoin(
+									queryCombinations(['stoner doom'],[globTags.genre, globTags.style],'OR'	),
+									'OR'
+								),
+								queryJoin(
+									queryCombinations(['folk metal'],[globTags.genre, globTags.style],'OR'	),
+									'OR'
+								),
+							], 'AND NOT'),
+							globTags.rating + ' IS 4 OR ' + globTags.rating + ' IS 3'
+						],'AND'),
+						_LIBRARY_3: queryJoin([
+							queryJoin([
+								queryJoin(
+									queryCombinations(['doom metal'],[globTags.genre, globTags.style],'OR'	),
+									'OR'
+								),
+								queryJoin(
+									queryCombinations(['folk metal'],[globTags.genre, globTags.style],'OR'	),
+									'OR'
+								),
+							], 'AND NOT'),
+							globTags.rating + ' IS 4 OR ' + globTags.rating + ' IS 3'
+						],'AND'),
+						_LIBRARY_4: queryJoin([
+							queryJoin([
+								queryJoin(
+									queryCombinations(['folk metal'],[globTags.genre, globTags.style],'OR'	),
+									'OR'
+								),
+								queryJoin(
+									queryCombinations(['black metal','stoner doom','doom metal','death metal'],[globTags.genre, globTags.style],'OR'	),
+									'OR'
+								),
+							], 'AND'),
+							globQuery.ratingGr2
+						],'AND'),
+						_LIBRARY_5: queryJoin([
+							queryJoin([
+								queryJoin(
+									queryCombinations(['black metal','stoner doom','doom metal','death metal'],[globTags.genre, globTags.style],'OR'	),
+									'OR'
+								),
+								queryJoin(
+									queryCombinations(['instrumental'],[globTags.genre, globTags.style],'OR'	),
+									'OR'
+								),
+							], 'AND'),
+							globQuery.ratingGr2
+						],'AND'),
+						_LIBRARY_6: queryJoin([
+							queryJoin([
+								queryJoin(
+									queryCombinations(['black metal','ambient metal'],[globTags.genre, globTags.style],'AND'),
+									'OR'
+								),
+								queryJoin(
+									queryCombinations(['death metal','acoustic'],[globTags.genre, globTags.style],'AND'	),
+									'OR'
+								),
+								queryJoin(
+									queryCombinations(['atmospheric black metal'],[globTags.genre, globTags.style],'OR'	),
+									'OR'
+								),
+							], 'OR'),
+							globQuery.ratingGr3
+						],'AND')
 					},
 					toPls: 'Dark Metal',
 					smartShuffle: 'ARTIST'
@@ -507,11 +595,11 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_4: tenth * 2
 					},
 					query: {
-						_LIBRARY_0: 'STYLE IS kawaii metal AND %RATING% IS 5',
-						_LIBRARY_1: 'STYLE IS j-pop AND (GENRE IS rock OR GENRE IS heavy metal) AND %RATING% GREATER 2',
-						_LIBRARY_2: 'STYLE IS kawaii metal AND (%RATING% IS 4 OR %RATING% IS 3)',
-						_LIBRARY_3: 'STYLE IS kawaii metal AND GENRE IS instrumental AND %RATING% GREATER 2',
-						_LIBRARY_4: 'STYLE IS kawaii metal OR (STYLE IS j-pop AND (GENRE IS rock OR GENRE IS heavy metal)) AND STYLE IS female vocal AND %RATING% GREATER 2'
+						_LIBRARY_0: globTags.style + ' IS kawaii metal AND ' + globQuery.fav,
+						_LIBRARY_1: globTags.style + ' IS j-pop AND (' + globTags.genre + ' IS rock OR ' + globTags.genre + ' IS heavy metal) AND ' + globQuery.ratingGr2,
+						_LIBRARY_2: globTags.style + ' IS kawaii metal AND (' + globTags.rating + ' IS 4 OR ' + globTags.rating + ' IS 3)',
+						_LIBRARY_3: globTags.style + ' IS kawaii metal AND ' + globTags.genre + ' IS instrumental AND ' + globQuery.ratingGr2,
+						_LIBRARY_4: globTags.style + ' IS kawaii metal OR (' + globTags.style + ' IS j-pop AND (' + globTags.genre + ' IS rock OR ' + globTags.genre + ' IS heavy metal)) AND ' + globTags.style + ' IS female vocal AND ' + globQuery.ratingGr2
 					},
 					toPls: 'Kawaii Metal',
 					smartShuffle: 'ARTIST'
@@ -530,12 +618,12 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_5: Math.floor(tenth * 0.9) || 1
 					},
 					query: {
-						_LIBRARY_0: '(STYLE IS chill-out downtempo OR STYLE IS trip hop OR STYLE IS deep house) AND %RATING% IS 5',
-						_LIBRARY_1: 'STYLE IS trip hop AND (MOOD IS not aggressive AND MOOD IS relaxed) AND (%RATING% IS 4 OR %RATING% IS 3)',
-						_LIBRARY_2: '(STYLE IS contemporary r&b OR STYLE IS neo soul) AND (MOOD IS not aggressive AND MOOD IS relaxed) AND (%RATING% IS 4 OR %RATING% IS 3)',
-						_LIBRARY_3: 'STYLE IS deep house AND (MOOD IS not aggressive AND MOOD IS relaxed AND BPM LESS 140) AND (%RATING% IS 4 OR %RATING% IS 3)',
-						_LIBRARY_4: '(STYLE IS chill-out downtempo OR STYLE IS trip hop OR STYLE IS deep house) AND STYLE IS female vocal AND (MOOD IS not aggressive AND MOOD IS relaxed) AND %RATING% GREATER 2',
-						_LIBRARY_5: '(STYLE IS chill-out downtempo OR STYLE IS trip hop OR STYLE IS deep house) AND GENRE IS instrumental AND (MOOD IS not aggressive AND MOOD IS relaxed) AND %RATING% GREATER 2'
+						_LIBRARY_0: '(' + globTags.style + ' IS chill-out downtempo OR ' + globTags.style + ' IS trip hop OR ' + globTags.style + ' IS deep house) AND ' + globQuery.fav,
+						_LIBRARY_1: globTags.style + ' IS trip hop AND (MOOD IS not aggressive AND MOOD IS relaxed) AND (' + globTags.rating + ' IS 4 OR ' + globTags.rating + ' IS 3)',
+						_LIBRARY_2: '(' + globTags.style + ' IS contemporary r&b OR ' + globTags.style + ' IS neo soul) AND (MOOD IS not aggressive AND MOOD IS relaxed) AND (' + globTags.rating + ' IS 4 OR ' + globTags.rating + ' IS 3)',
+						_LIBRARY_3: globTags.style + ' IS deep house AND (MOOD IS not aggressive AND MOOD IS relaxed AND BPM LESS 140) AND (' + globTags.rating + ' IS 4 OR ' + globTags.rating + ' IS 3)',
+						_LIBRARY_4: '(' + globTags.style + ' IS chill-out downtempo OR ' + globTags.style + ' IS trip hop OR ' + globTags.style + ' IS deep house) AND ' + globTags.style + ' IS female vocal AND (MOOD IS not aggressive AND MOOD IS relaxed) AND ' + globQuery.ratingGr2,
+						_LIBRARY_5: '(' + globTags.style + ' IS chill-out downtempo OR ' + globTags.style + ' IS trip hop OR ' + globTags.style + ' IS deep house) AND ' + globTags.genre + ' IS instrumental AND (MOOD IS not aggressive AND MOOD IS relaxed) AND ' + globQuery.ratingGr2
 					},
 					toPls: 'Chill and Downtempo',
 					smartShuffle: 'ARTIST'
@@ -557,22 +645,22 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_8: Math.floor(tenth * 1.4)
 					},
 					query: {
-						_LIBRARY_0: '(STYLE IS trap OR GENRE IS hip-hop) AND (LANGUAGE IS spa OR STYLE IS spanish hip-hop) AND NOT (STYLE IS rap metal) AND %RATING% IS 5',
-						_LIBRARY_1: '(STYLE IS trap OR GENRE IS hip-hop) AND (LANGUAGE IS spa OR STYLE IS spanish hip-hop OR STYLE IS latin trap) AND NOT (STYLE IS rap metal) AND (%RATING% IS 4 OR %RATING% IS 3)',
-						_LIBRARY_2: 'STYLE IS spanish hip-hop AND NOT (STYLE IS rap metal) AND (%RATING% IS 4 OR %RATING% IS 3)',
-						_LIBRARY_3: '(STYLE IS flamenco OR STYLE IS rumba flamenca) AND GENRE IS hip-hop AND %RATING% GREATER 2',
-						_LIBRARY_4: '(STYLE IS trap OR GENRE IS hip-hop) AND (LANGUAGE IS spa OR STYLE IS spanish hip-hop OR STYLE IS latin trap) AND NOT (STYLE IS rap metal) AND STYLE IS female vocal AND %RATING% GREATER 2',
-						_LIBRARY_5: 'STYLE IS nuevo flamenco AND GENRE IS hip-hop AND DATE GREATER 2000 AND %RATING% GREATER 3',
-						_LIBRARY_6: '(STYLE IS flamenco rock OR STYLE IS flamenco OR STYLE IS rumba flamenca) AND (STYLE IS spanish folk OR STYLE IS spanish rock) AND DATE GREATER 2000 AND %RATING% GREATER 3',
-						_LIBRARY_7: '(STYLE IS rumba fusion OR STYLE IS rumba) AND STYLE IS spanish rock AND DATE GREATER 2000 AND %RATING% GREATER 3',
-						_LIBRARY_8: '(STYLE IS trap OR GENRE IS hip-hop) AND (LANGUAGE IS spa OR STYLE IS spanish hip-hop OR STYLE IS latin trap) AND NOT (STYLE IS rap metal) AND %RATING% IS 5',
+						_LIBRARY_0: '(' + globTags.style + ' IS trap OR ' + globTags.genre + ' IS hip-hop) AND (LANGUAGE IS spa OR ' + globTags.style + ' IS spanish hip-hop) AND NOT (' + globTags.style + ' IS rap metal) AND ' + globQuery.fav,
+						_LIBRARY_1: '(' + globTags.style + ' IS trap OR ' + globTags.genre + ' IS hip-hop) AND (LANGUAGE IS spa OR ' + globTags.style + ' IS spanish hip-hop OR ' + globTags.style + ' IS latin trap) AND NOT (' + globTags.style + ' IS rap metal) AND (' + globTags.rating + ' IS 4 OR ' + globTags.rating + ' IS 3)',
+						_LIBRARY_2: globTags.style + ' IS spanish hip-hop AND NOT (' + globTags.style + ' IS rap metal) AND (' + globTags.rating + ' IS 4 OR ' + globTags.rating + ' IS 3)',
+						_LIBRARY_3: '(' + globTags.style + ' IS flamenco OR ' + globTags.style + ' IS rumba flamenca) AND ' + globTags.genre + ' IS hip-hop AND ' + globQuery.ratingGr2,
+						_LIBRARY_4: '(' + globTags.style + ' IS trap OR ' + globTags.genre + ' IS hip-hop) AND (LANGUAGE IS spa OR ' + globTags.style + ' IS spanish hip-hop OR ' + globTags.style + ' IS latin trap) AND NOT (' + globTags.style + ' IS rap metal) AND ' + globTags.style + ' IS female vocal AND ' + globQuery.ratingGr2,
+						_LIBRARY_5: globTags.style + ' IS nuevo flamenco AND ' + globTags.genre + ' IS hip-hop AND DATE GREATER 2000 AND ' + globQuery.ratingGr3,
+						_LIBRARY_6: '(' + globTags.style + ' IS flamenco rock OR ' + globTags.style + ' IS flamenco OR ' + globTags.style + ' IS rumba flamenca) AND (' + globTags.style + ' IS spanish folk OR ' + globTags.style + ' IS spanish rock) AND DATE GREATER 2000 AND ' + globQuery.ratingGr3,
+						_LIBRARY_7: '(' + globTags.style + ' IS rumba fusion OR ' + globTags.style + ' IS rumba) AND ' + globTags.style + ' IS spanish rock AND DATE GREATER 2000 AND ' + globQuery.ratingGr3,
+						_LIBRARY_8: '(' + globTags.style + ' IS trap OR ' + globTags.genre + ' IS hip-hop) AND (LANGUAGE IS spa OR ' + globTags.style + ' IS spanish hip-hop OR ' + globTags.style + ' IS latin trap) AND NOT (' + globTags.style + ' IS rap metal) AND ' + globQuery.fav,
 					},
 					toPls: 'Spanish Urban Music',
 					smartShuffle: 'ARTIST'
 				}
 			},
 			{
-				name: 'Rock Classics',
+				name: 'Rock Classics (until 90s)',
 				folder: 'Genre mixes',
 				pool: {
 					fromPls: {
@@ -587,17 +675,47 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_8: Math.floor(tenth * 0.5) || 1,
 					},
 					query: {
-						_LIBRARY_0: 'STYLE IS classic rock AND GENRE IS rock AND DATE LESS 1990 AND %RATING% IS 5',
-						_LIBRARY_1: 'STYLE IS classic rock AND GENRE IS rock AND DATE LESS 1990 AND %RATING% IS 4',
-						_LIBRARY_2: 'STYLE IS classic rock AND GENRE IS rock AND DATE LESS 1990 AND %RATING% IS 3',
-						_LIBRARY_3: 'STYLE IS classic rock AND GENRE IS folk-rock AND DATE LESS 1990 AND %RATING% GREATER 3',
-						_LIBRARY_4: 'STYLE IS classic rock AND STYLE IS female vocal AND DATE LESS 1990 AND %RATING% GREATER 2',
-						_LIBRARY_5: 'GENRE IS rock AND NOT STYLE IS 80s rock AND DATE GREATER 1960 AND DATE LESS 1980 AND %RATING% GREATER 3',
-						_LIBRARY_6: 'GENRE IS hard rock AND STYLE IS classic rock AND DATE LESS 1990 AND %RATING% GREATER 2',
-						_LIBRARY_7: '(STYLE IS beat music OR STYLE IS soft rock OR GENRE IS rock & roll) AND DATE GREATER 1960 AND DATE LESS 1980 AND %RATING% IS 5',
-						_LIBRARY_8: 'GENRE IS rock & roll AND DATE GREATER 1960 AND DATE LESS 1980 AND %RATING% IS 5',
+						_LIBRARY_0: globTags.style + ' IS classic rock AND ' + globTags.genre + ' IS rock AND DATE LESS 1990 AND ' + globQuery.fav,
+						_LIBRARY_1: globTags.style + ' IS classic rock AND ' + globTags.genre + ' IS rock AND DATE LESS 1990 AND ' + globTags.rating + ' IS 4',
+						_LIBRARY_2: globTags.style + ' IS classic rock AND ' + globTags.genre + ' IS rock AND DATE LESS 1990 AND ' + globTags.rating + ' IS 3',
+						_LIBRARY_3: globTags.style + ' IS classic rock AND ' + globTags.genre + ' IS folk-rock AND DATE LESS 1990 AND ' + globQuery.ratingGr3,
+						_LIBRARY_4: globTags.style + ' IS classic rock AND ' + globTags.style + ' IS female vocal AND DATE LESS 1990 AND ' + globQuery.ratingGr2,
+						_LIBRARY_5: globTags.genre + ' IS rock AND NOT ' + globTags.style + ' IS 80s rock AND DATE GREATER 1960 AND DATE LESS 1980 AND ' + globQuery.ratingGr3,
+						_LIBRARY_6: globTags.genre + ' IS hard rock AND ' + globTags.style + ' IS classic rock AND DATE LESS 1990 AND ' + globQuery.ratingGr2,
+						_LIBRARY_7: '(' + globTags.style + ' IS beat music OR ' + globTags.style + ' IS soft rock OR ' + globTags.genre + ' IS rock & roll) AND DATE GREATER 1960 AND DATE LESS 1980 AND ' + globQuery.fav,
+						_LIBRARY_8: globTags.genre + ' IS rock & roll AND DATE GREATER 1960 AND DATE LESS 1980 AND ' + globQuery.fav,
 					},
-					toPls: 'Rock Classics',
+					toPls: 'Rock Classics (until 90s)',
+					smartShuffle: 'ARTIST'
+				}
+			},
+			{
+				name: '60s & 70s Rock Classics',
+				folder: 'Genre mixes',
+				pool: {
+					fromPls: {
+						_LIBRARY_0: Math.floor(tenth * 1.4),
+						_LIBRARY_1: Math.floor(tenth * 1.5),
+						_LIBRARY_2: Math.floor(tenth * 1.4),
+						_LIBRARY_3: Math.floor(tenth * 0.8) || 1,
+						_LIBRARY_4: tenth,
+						_LIBRARY_5: Math.floor(tenth * 1.4) || 1,
+						_LIBRARY_6: tenth,
+						_LIBRARY_7: Math.floor(tenth * 0.8) || 1,
+						_LIBRARY_8: Math.floor(tenth * 0.5) || 1,
+					},
+					query: {
+						_LIBRARY_0: globTags.style + ' IS classic rock AND ' + globTags.genre + ' IS rock AND DATE LESS 1979 AND ' + globQuery.fav,
+						_LIBRARY_1: globTags.style + ' IS classic rock AND ' + globTags.genre + ' IS rock AND DATE LESS 1979 AND ' + globTags.rating + ' IS 4',
+						_LIBRARY_2: globTags.style + ' IS classic rock AND ' + globTags.genre + ' IS rock AND DATE LESS 1979 AND ' + globTags.rating + ' IS 3',
+						_LIBRARY_3: globTags.style + ' IS classic rock AND ' + globTags.genre + ' IS folk-rock AND DATE LESS 1979 AND ' + globQuery.ratingGr3,
+						_LIBRARY_4: globTags.style + ' IS classic rock AND ' + globTags.style + ' IS female vocal AND DATE LESS 1979 AND ' + globQuery.ratingGr2,
+						_LIBRARY_5: globTags.genre + ' IS rock AND NOT ' + globTags.style + ' IS 80s rock AND DATE GREATER 1959 AND DATE LESS 1979 AND ' + globQuery.ratingGr3,
+						_LIBRARY_6: globTags.genre + ' IS hard rock AND ' + globTags.style + ' IS classic rock AND DATE LESS 1979 AND ' + globQuery.ratingGr2,
+						_LIBRARY_7: '(' + globTags.style + ' IS beat music OR ' + globTags.style + ' IS soft rock OR ' + globTags.genre + ' IS rock & roll) AND DATE GREATER 1959 AND DATE LESS 1979 AND ' + globQuery.fav,
+						_LIBRARY_8: globTags.genre + ' IS rock & roll AND DATE GREATER 1959 AND DATE LESS 1979 AND ' + globQuery.fav,
+					},
+					toPls: '60s & 70s Rock Classics',
 					smartShuffle: 'ARTIST'
 				}
 			},
@@ -617,15 +735,15 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_8: tenth,
 					},
 					query: {
-						_LIBRARY_0: 'STYLE IS acid rock AND GENRE IS psychedelic rock AND DATE GREATER 1960 AND DATE LESS 1971 AND %RATING% IS 5',
-						_LIBRARY_1: '(STYLE IS acid rock OR STYLE IS raga rock) AND GENRE IS psychedelic rock AND DATE GREATER 1960 AND DATE LESS 1971 AND %RATING% GREATER 3',
-						_LIBRARY_2: 'STYLE IS acid rock AND (GENRE IS rock OR GENRE IS folk-rock) AND DATE GREATER 1960 AND DATE LESS 1971 AND %RATING% GREATER 2',
-						_LIBRARY_3: 'STYLE IS british psychedelia AND GENRE IS psychedelic rock AND DATE GREATER 1960 AND DATE LESS 1971 AND %RATING% IS 5',
-						_LIBRARY_4: 'STYLE IS british psychedelia AND (GENRE IS rock OR GENRE IS folk-rock OR GENRE IS psychedelic rock) AND DATE GREATER 1960 AND DATE LESS 1971 AND %RATING% GREATER 2',
-						_LIBRARY_5: '(STYLE IS krautrock OR STYLE IS space rock OR STYLE IS italian prog. rock OR STYLE IS japanese prog. rock) AND GENRE IS psychedelic rock AND DATE GREATER 1960 AND DATE LESS 1971 AND %RATING% GREATER 3',
-						_LIBRARY_6: 'STYLE IS psychedelic folk AND DATE GREATER 1960 AND DATE LESS 1971 AND %RATING% GREATER 2',
-						_LIBRARY_7: '(STYLE IS psychedelic pop OR STYLE IS sunshine pop) AND DATE GREATER 1960 AND DATE LESS 1971 AND %RATING% GREATER 3',
-						_LIBRARY_8: '(STYLE IS psychedelic soul OR STYLE IS psychedelic funk OR STYLE IS psychedelic blues) AND DATE GREATER 1960 AND DATE LESS 1971 AND %RATING% IS 5',
+						_LIBRARY_0: globTags.style + ' IS acid rock AND ' + globTags.genre + ' IS psychedelic rock AND DATE GREATER 1960 AND DATE LESS 1971 AND ' + globQuery.fav,
+						_LIBRARY_1: '(' + globTags.style + ' IS acid rock OR ' + globTags.style + ' IS raga rock) AND ' + globTags.genre + ' IS psychedelic rock AND DATE GREATER 1960 AND DATE LESS 1971 AND ' + globQuery.ratingGr3,
+						_LIBRARY_2: globTags.style + ' IS acid rock AND (' + globTags.genre + ' IS rock OR ' + globTags.genre + ' IS folk-rock) AND DATE GREATER 1960 AND DATE LESS 1971 AND ' + globQuery.ratingGr2,
+						_LIBRARY_3: globTags.style + ' IS british psychedelia AND ' + globTags.genre + ' IS psychedelic rock AND DATE GREATER 1960 AND DATE LESS 1971 AND ' + globQuery.fav,
+						_LIBRARY_4: globTags.style + ' IS british psychedelia AND (' + globTags.genre + ' IS rock OR ' + globTags.genre + ' IS folk-rock OR ' + globTags.genre + ' IS psychedelic rock) AND DATE GREATER 1960 AND DATE LESS 1971 AND ' + globQuery.ratingGr2,
+						_LIBRARY_5: '(' + globTags.style + ' IS krautrock OR ' + globTags.style + ' IS space rock OR ' + globTags.style + ' IS italian prog. rock OR ' + globTags.style + ' IS japanese prog. rock) AND ' + globTags.genre + ' IS psychedelic rock AND DATE GREATER 1960 AND DATE LESS 1971 AND ' + globQuery.ratingGr3,
+						_LIBRARY_6: globTags.style + ' IS psychedelic folk AND DATE GREATER 1960 AND DATE LESS 1971 AND ' + globQuery.ratingGr2,
+						_LIBRARY_7: '(' + globTags.style + ' IS psychedelic pop OR ' + globTags.style + ' IS sunshine pop) AND DATE GREATER 1960 AND DATE LESS 1971 AND ' + globQuery.ratingGr3,
+						_LIBRARY_8: '(' + globTags.style + ' IS psychedelic soul OR ' + globTags.style + ' IS psychedelic funk OR ' + globTags.style + ' IS psychedelic blues) AND DATE GREATER 1960 AND DATE LESS 1971 AND ' + globQuery.fav,
 					},
 					toPls: '60s Psychedelic Pills',
 					smartShuffle: 'ARTIST'
@@ -647,15 +765,15 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_8: Math.floor(tenth * 0.7) || 1,
 					},
 					query: {
-						_LIBRARY_0: 'STYLE IS 80s rock AND GENRE IS rock AND DATE LESS 1992 AND %RATING% IS 5',
-						_LIBRARY_1: 'STYLE IS 80s rock AND (GENRE IS rock OR GENRE IS pop) AND DATE LESS 1992 AND %RATING% IS 4',
-						_LIBRARY_2: 'STYLE IS 80s rock AND (GENRE IS rock OR GENRE IS pop) AND DATE LESS 1992 AND %RATING% IS 3',
-						_LIBRARY_3: 'STYLE IS 80s rock AND (GENRE IS folk-rock OR GENRE IS blues OR GENRE IS alt. rock) AND DATE LESS 1992 AND %RATING% GREATER 3',
-						_LIBRARY_4: 'STYLE IS 80s rock OR (GENRE IS rock AND DATE GREATER 1979 AND DATE LESS 1990) AND STYLE IS female vocal AND %RATING% GREATER 2',
-						_LIBRARY_5: 'STYLE IS power pop AND DATE GREATER 1979 AND DATE LESS 1990 AND %RATING% GREATER 3',
-						_LIBRARY_6: 'STYLE IS new wave AND DATE GREATER 1979 AND DATE LESS 1990 AND %RATING% GREATER 3',
-						_LIBRARY_7: '(STYLE IS post-punk OR STYLE IS soft rock OR GENRE IS rock & roll) AND DATE GREATER 1979 AND DATE LESS 1990 AND %RATING% GREATER 3',
-						_LIBRARY_8: '(STYLE IS euro-pop OR STYLE IS sophisti-pop OR STYLE IS disco OR STYLE IS smooth soul) AND DATE GREATER 1979 AND DATE LESS 1990 AND %RATING% GREATER 3',
+						_LIBRARY_0: globTags.style + ' IS 80s rock AND ' + globTags.genre + ' IS rock AND DATE LESS 1992 AND ' + globQuery.fav,
+						_LIBRARY_1: globTags.style + ' IS 80s rock AND (' + globTags.genre + ' IS rock OR ' + globTags.genre + ' IS pop) AND DATE LESS 1992 AND ' + globTags.rating + ' IS 4',
+						_LIBRARY_2: globTags.style + ' IS 80s rock AND (' + globTags.genre + ' IS rock OR ' + globTags.genre + ' IS pop) AND DATE LESS 1992 AND ' + globTags.rating + ' IS 3',
+						_LIBRARY_3: globTags.style + ' IS 80s rock AND (' + globTags.genre + ' IS folk-rock OR ' + globTags.genre + ' IS blues OR ' + globTags.genre + ' IS alt. rock) AND DATE LESS 1992 AND ' + globQuery.ratingGr3,
+						_LIBRARY_4: globTags.style + ' IS 80s rock OR (' + globTags.genre + ' IS rock AND DATE GREATER 1979 AND DATE LESS 1990) AND ' + globTags.style + ' IS female vocal AND ' + globQuery.ratingGr2,
+						_LIBRARY_5: globTags.style + ' IS power pop AND DATE GREATER 1979 AND DATE LESS 1990 AND ' + globQuery.ratingGr3,
+						_LIBRARY_6: globTags.style + ' IS new wave AND DATE GREATER 1979 AND DATE LESS 1990 AND ' + globQuery.ratingGr3,
+						_LIBRARY_7: '(' + globTags.style + ' IS post-punk OR ' + globTags.style + ' IS soft rock OR ' + globTags.genre + ' IS rock & roll) AND DATE GREATER 1979 AND DATE LESS 1990 AND ' + globQuery.ratingGr3,
+						_LIBRARY_8: '(' + globTags.style + ' IS euro-pop OR ' + globTags.style + ' IS sophisti-pop OR ' + globTags.style + ' IS disco OR ' + globTags.style + ' IS smooth soul) AND DATE GREATER 1979 AND DATE LESS 1990 AND ' + globQuery.ratingGr3,
 					},
 					toPls: '80s Hits',
 					smartShuffle: 'ARTIST'
@@ -674,12 +792,12 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_5: Math.floor(tenth * 0.9) || 1
 					},
 					query: {
-						_LIBRARY_0: '(STYLE IS darkwave OR STYLE IS gothic rock OR STYLE IS post-punk) AND %RATING% IS 5',
-						_LIBRARY_1: '(STYLE IS darkwave OR STYLE IS gothic rock OR STYLE IS post-punk) AND DATE LESS 2000 AND %RATING% IS 5',
-						_LIBRARY_2: '(STYLE IS darkwave OR STYLE IS gothic rock OR STYLE IS post-punk) AND %RATING% GREATER 2',
-						_LIBRARY_3: '(STYLE IS darkwave OR STYLE IS gothic rock OR STYLE IS post-punk) AND STYLE IS female vocal AND %RATING% GREATER 2',
-						_LIBRARY_4: '(STYLE IS gothic metal) AND %RATING% GREATER 3',
-						_LIBRARY_5: '(STYLE IS darksynth OR STYLE IS dark techno) AND %RATING% GREATER 3'
+						_LIBRARY_0: '(' + globTags.style + ' IS darkwave OR ' + globTags.style + ' IS gothic rock OR ' + globTags.style + ' IS post-punk) AND ' + globQuery.fav,
+						_LIBRARY_1: '(' + globTags.style + ' IS darkwave OR ' + globTags.style + ' IS gothic rock OR ' + globTags.style + ' IS post-punk) AND DATE LESS 2000 AND ' + globQuery.fav,
+						_LIBRARY_2: '(' + globTags.style + ' IS darkwave OR ' + globTags.style + ' IS gothic rock OR ' + globTags.style + ' IS post-punk) AND ' + globQuery.ratingGr2,
+						_LIBRARY_3: '(' + globTags.style + ' IS darkwave OR ' + globTags.style + ' IS gothic rock OR ' + globTags.style + ' IS post-punk) AND ' + globTags.style + ' IS female vocal AND ' + globQuery.ratingGr2,
+						_LIBRARY_4: '(' + globTags.style + ' IS gothic metal) AND ' + globQuery.ratingGr3,
+						_LIBRARY_5: '(' + globTags.style + ' IS darksynth OR ' + globTags.style + ' IS dark techno) AND ' + globQuery.ratingGr3
 					},
 					toPls: 'Gothic Rock',
 					smartShuffle: 'ARTIST'
@@ -701,20 +819,207 @@ function createPoolPresets({size = 50} = {}) {
 						_LIBRARY_8: tenth
 					},
 					query: {
-						_LIBRARY_0: 'STYLE IS gothic metal AND %RATING% IS 5',
-						_LIBRARY_1: 'STYLE IS atmospheric black metal AND %RATING% GREATER 3',
-						_LIBRARY_2: 'STYLE IS black metal AND %RATING% IS 5',
-						_LIBRARY_3: '(STYLE IS dark techno OR STYLE IS dark ambient) AND GENRE IS heavy metal AND %RATING% GREATER 3',
-						_LIBRARY_4: 'STYLE IS doom metal AND GENRE IS heavy metal AND %RATING% GREATER 3',
-						_LIBRARY_5: 'STYLE IS doom metal AND GENRE IS heavy metal AND STYLE IS female vocal AND %RATING% GREATER 3',
-						_LIBRARY_6: 'STYLE IS doom metal AND (%RATING% IS 5 OR STYLE IS female vocal AND %RATING% IS 4)',
-						_LIBRARY_7: '(STYLE IS pagan metal OR STYLE IS sludge metal OR STYLE IS stoner doom OR STYLE IS stoner sludge) AND GENRE IS heavy metal AND %RATING% GREATER 3',
-						_LIBRARY_8: 'STYLE IS gothic metal AND STYLE IS female vocal AND %RATING% IS 5'
+						_LIBRARY_0: globTags.style + ' IS gothic metal AND ' + globQuery.fav,
+						_LIBRARY_1: globTags.style + ' IS atmospheric black metal AND ' + globQuery.ratingGr3,
+						_LIBRARY_2: globTags.style + ' IS black metal AND ' + globQuery.fav,
+						_LIBRARY_3: '(' + globTags.style + ' IS dark techno OR ' + globTags.style + ' IS dark ambient) AND ' + globTags.genre + ' IS heavy metal AND ' + globQuery.ratingGr3,
+						_LIBRARY_4: globTags.style + ' IS doom metal AND ' + globTags.genre + ' IS heavy metal AND ' + globQuery.ratingGr3,
+						_LIBRARY_5: globTags.style + ' IS doom metal AND ' + globTags.genre + ' IS heavy metal AND ' + globTags.style + ' IS female vocal AND ' + globQuery.ratingGr3,
+						_LIBRARY_6: globTags.style + ' IS doom metal AND (' + globTags.rating + ' IS 5 OR ' + globTags.style + ' IS female vocal AND ' + globTags.rating + ' IS 4)',
+						_LIBRARY_7: '(' + globTags.style + ' IS pagan metal OR ' + globTags.style + ' IS sludge metal OR ' + globTags.style + ' IS stoner doom OR ' + globTags.style + ' IS stoner sludge) AND ' + globTags.genre + ' IS heavy metal AND ' + globQuery.ratingGr3,
+						_LIBRARY_8: globTags.style + ' IS gothic metal AND ' + globTags.style + ' IS female vocal AND ' + globQuery.fav
 					},
 					toPls: 'Gothic Rock',
 					smartShuffle: 'ARTIST'
 				}
+			},
+			{
+				name: 'Acoustic Bal Folk',
+				folder: 'Genre mixes',
+				pool: {
+					fromPls: {
+						_LIBRARY_0: fourth,
+						_LIBRARY_1: fourth,
+						_LIBRARY_2: fourth,
+						_LIBRARY_3: fourth,
+					},
+					query: {
+						_LIBRARY_0: queryJoin(
+							[
+								queryJoin(
+									queryCombinations(
+										['andro', 'bourree', 'bresse', 'chapelloise', 'circle', 'farelquesh', 'gavotte', 'hanterdro', 'kost ar c\'hoad', 'laride', 'mazurka', 'jig', 'plinn', 'polka', 'rond', 'scottish', 'tarantella', 'tricot', 'vals', 'bal folk', 'traditional european folk'],
+										[globTags.genre, globTags.style],
+										'OR'
+									),
+									'OR'
+								),
+								queryJoin(queryCombinations(['folk'], [globTags.genre, globTags.style], 'OR'), 'OR'),
+								globQuery.fav
+							],
+							'AND'
+						),
+						_LIBRARY_1: queryJoin(
+							[
+								queryJoin(
+									queryCombinations(
+										['andro', 'bourree', 'bresse', 'chapelloise', 'circle', 'farelquesh', 'gavotte', 'hanterdro', 'kost ar c\'hoad', 'laride', 'mazurka', 'jig', 'plinn', 'polka', 'rond', 'scottish', 'tarantella', 'tricot', 'vals', 'bal folk', 'traditional european folk'],
+										[globTags.genre, globTags.style],
+										'OR'
+									),
+									'OR'
+								),
+								queryJoin(queryCombinations(['folk'], [globTags.genre, globTags.style], 'OR'), 'OR'),
+								globQuery.ratingGr3
+							],
+							'AND'
+						),
+						_LIBRARY_2: queryJoin(
+							[
+								queryJoin(
+									queryCombinations(
+										['andro', 'bourree', 'bresse', 'chapelloise', 'circle', 'farelquesh', 'gavotte', 'hanterdro', 'kost ar c\'hoad', 'laride', 'mazurka', 'jig', 'plinn', 'polka', 'rond', 'scottish', 'tarantella', 'tricot', 'vals', 'bal folk', 'traditional european folk'],
+										[globTags.genre, globTags.style],
+										'OR'
+									),
+									'OR'
+								),
+								queryJoin(queryCombinations(['folk'], [globTags.genre, globTags.style], 'OR'), 'OR'),
+								queryJoin(queryCombinations(['instrumental'], [globTags.genre, globTags.style], 'OR'), 'OR'),
+								globQuery.ratingGr3
+							],
+							'AND'
+						),
+						_LIBRARY_3: queryJoin(
+							[
+								queryJoin(
+									queryCombinations(
+										['traditional european folk'],
+										[globTags.genre, globTags.style],
+										'OR'
+									),
+									'OR'
+								),
+								queryJoin(queryCombinations(['folk'], [globTags.genre, globTags.style], 'OR'), 'OR'),
+								globQuery.ratingGr2
+							],
+							'AND'
+						)
+					},
+					toPls: 'Acoustic Bal Folk',
+					smartShuffle: 'ARTIST'
+				}
+			},
+			{
+				name: 'Acoustic for Reading',
+				folder: 'Genre mixes',
+				pool: {
+					fromPls: {
+						_LIBRARY_0: fourth,
+						_LIBRARY_1: fourth,
+						_LIBRARY_2: eighth,
+						_LIBRARY_3: eighth,
+						_LIBRARY_4: eighth,
+						_LIBRARY_5: eighth,
+					},
+					query: {
+						_LIBRARY_0: queryJoin(
+							[
+								queryJoin(
+									queryCombinations(
+										['traditional european folk', 'ambient classical', 'ambient folk', 'ambient new age', 'american primitive guitar', 'neo-classical new age', 'new acoustic'],
+										[globTags.genre, globTags.style],
+										'OR'
+									),
+									'OR'
+								),
+								queryJoin(queryCombinations(['folk', 'new age', 'classical', 'acoustic'], [globTags.genre, globTags.style], 'OR'), 'OR'),
+								globQuery.fav
+							],
+							'AND'
+						),
+						_LIBRARY_1: queryJoin(
+							[
+								queryJoin(
+									queryCombinations(
+										['ambient classical', 'neo-classical new age', 'new acoustic'],
+										[globTags.genre, globTags.style],
+										'OR'
+									),
+									'OR'
+								),
+								queryJoin(queryCombinations(['folk', 'new age', 'classical', 'acoustic'], [globTags.genre, globTags.style], 'OR'), 'OR'),
+								globQuery.ratingGr3
+							],
+							'AND'
+						),
+						_LIBRARY_2: queryJoin(
+							[
+								queryJoin(
+									queryCombinations(
+										['british pychedelia', 'americana', 'asian folk', 'hang music', 'healing music', 'appalachian', 'psychedelic folk', 'nubian folk'],
+										[globTags.genre, globTags.style],
+										'OR'
+									),
+									'OR'
+								),
+								queryJoin(queryCombinations(['folk', 'new age', 'classical', 'acoustic', 'ambient'], [globTags.genre, globTags.style], 'OR'), 'OR'),
+								globQuery.ratingGr3
+							],
+							'AND'
+						),
+						_LIBRARY_3: queryJoin(
+							[
+								queryJoin(
+									queryCombinations(
+										['british pychedelia', 'americana', 'asian folk', 'hang music', 'healing music', 'appalachian', 'psychedelic folk'],
+										[globTags.genre, globTags.style],
+										'OR'
+									),
+									'OR'
+								),
+								queryJoin(queryCombinations(['folk', 'new age', 'classical', 'acoustic', 'ambient'], [globTags.genre, globTags.style], 'OR'), 'OR'),
+								queryJoin(queryCombinations(['instrumental'], [globTags.genre, globTags.style], 'OR'), 'OR'),
+								globQuery.ratingGr3
+							],
+							'AND'
+						),
+						_LIBRARY_4: queryJoin(
+							[
+								queryJoin(
+									queryCombinations(
+										['folk metal', 'Atmospheric Black Metal'],
+										[globTags.genre, globTags.style],
+										'OR'
+									),
+									'OR'
+								),
+								queryJoin(queryCombinations(['instrumental', 'acoustic', 'folk'], [globTags.genre, globTags.style], 'OR'), 'OR'),
+								globQuery.ratingGr3
+							],
+							'AND'
+						),
+						_LIBRARY_5: queryJoin(
+							[
+								queryJoin(
+									queryCombinations(
+										['celtic new age', 'tuvan', 'tuareg music', 'traditional european folk', 'sephardic', 'pagan folk'],
+										[globTags.genre, globTags.style],
+										'OR'
+									),
+									'OR'
+								),
+								queryJoin(queryCombinations(['world'], [globTags.genre, globTags.style], 'OR'), 'OR'),
+								queryJoin(queryCombinations(['instrumental', 'acoustic', 'ambient folk', 'folk'], [globTags.genre, globTags.style], 'OR'), 'OR'),
+								globQuery.ratingGr3
+							],
+							'AND'
+						)
+					},
+					toPls: 'Acoustic for Reading',
+					smartShuffle: 'ARTIST'
+				}
 			}
-		]
+		].sort((a, b) => a.name.localeCompare(b.name))
 	];
 }